add component and bind decorators
