+++ rollup.config.js
@@ -3,7 +3,7 @@ import typescript from "rollup-plugin-typescript2";
 import pkg from "./package.json";
 
 export default {
-  input: "src/aframe-typescript-class-components.ts",
+  input: "src/index.ts",
   output: [
     {
       file: pkg.main,
