+++ src/index.ts
@@ -1,0 +1,5 @@
+export * from "BaseComponent";
+export * from "toComponent";
+export * from "decorators";
+export * from "decorators";
+export * from "test-helpers";
